moved predictor to library
